fix: multihashes does not export default lengths any more
Since https://github.com/multiformats/js-multihash/pull/76 was released
the `multihashes` module does not export default hash lengths any more.

For reasons that are unclear it went out as a patch release which means
`ipld-git` is currently broken.

This PR hard codes the hash length as it is known, and also removes
the `multihashes` dep as we can use the one exported from `multihashing-async`
and updates other deps.

Ref: https://github.com/multiformats/js-multihash/pull/76#issuecomment-646561123